[Refactor ⚙️] ProfileBar 컴포넌트 분리
