+++ src/pages/ProfilePage/components/Profile/ProfileBarView.tsx
